refactor: reorder imports, core types, skillsBadgesList and ExternalDomainFavicon
