+++ components/core/certificates/index.tsx
@@ -1,2 +1,2 @@
-export { CertificateCard } from '@/components/core/certificates/card'
-export { CertificatesPanel } from '@/components/core/certificates/panel'
+export { CertificateCard } from './card'
+export { CertificatesPanel } from './panel'
