+++ components/core/certificates/panel.tsx
@@ -1,5 +1,6 @@
-import { CertificateCard } from '@/components/core/certificates'
 import { Certificate } from '@/types/core'
+
+import { CertificateCard } from './index'
 
 
 interface CertificatesPanelProps {
