+++ components/core/skills/index.tsx
@@ -1,3 +1,2 @@
-export { SkillBadge } from '@/components/core/skills/badge'
-export { SkillBadgesList } from '@/components/core/skills/badge-list'
+export { SkillBadge, SkillBadgesList } from '@/components/core/skills/badge-list'
 export type { ToggleSkill, SkillIsActive } from '@/components/core/skills/types'
