+++ types/core/index.ts
@@ -1,3 +1,19 @@
-export type { Certificate } from '@/types/core/certificates'
-export type { Reference } from '@/types/core/references'
-export type { Skill } from '@/types/core/skills'
+import { Raindrop } from '@/types/data/raindrops'
+
+
+export interface Certificate {
+    title: string,
+    date: string, // TODO: define date type (mm/dd/aaaa | mm/aaaa)
+    skills: string[],
+    issuer: {
+            title: string,
+            logo: string
+    },
+    imageUrl: string,
+    certificateUrl: string
+}
+
+export type Reference = Raindrop
+
+export type Skill = string
+
