--- types/core/references.ts
@@ -1,4 +1,0 @@
-import { Raindrop } from '@/types/data/raindrops'
-
-
-export type Reference = Raindrop
