--- types/core/skills.ts
@@ -1,1 +1,0 @@
-export type Skill = string
